embedder/requirements.txt
kafka-python
sentence-transformers
pymilvus
numpy
qdrant-client